# backend/src/main/resources/static/swagger.yaml
openapi: 3.0.3
info:
  title: TechFixApi REST
  version: 2.0.0
  description: |
    A TechFixApi é uma aplicação desenvolvida em Java e Spring Boot para gerenciar operações de assistência técnica. O sistema processa o controle de estoque de peças e o catálogo de serviços. A API gerencia o acesso de usuários através de tokens JWT e perfis de permissão. As respostas de erro seguem um padrão técnico, e os dados são entregues com paginação. Este é um projeto pessoal criado para aplicar e testar meus conhecimentos em desenvolvimento backend
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: "http://localhost:8080"
    description: "Servidor Teste"
security:
  - bearerAuth: []
paths:
...
  /api/v2/service-orders/{id}:
    parameters: 
      - name: id
        in: path  
        schema:
          type: number
        required: true
        description: ID Solicitação de Serviço
    get:
      tags: 
        - Service Orders
      summary: Busca detalhes da Ordem Serviço pelo ID
      description: Endpoint realiza consulta de Ordem Serviço pelo ID.Traz informações do Cliente e da Solicitação de Serviço com datas e status para melhor acompanhamento.
      responses:
        '200':
          description: Busca realizada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrderDetailDTO'
        '403':
            description: Acesso negado. Apenas usuários com permissão.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetail'
        '404':
            description: Ordem de Serviço não encontrada com ID informado.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetail'
    
    put: 
      tags:
        - Service Orders
      summary: Atualiza Ordem de Serviço
      description: |
       Endpoint atualiza Ordem de Serviço permitindo atualizar status e atribuir técnico a OS.
       
       ### Regras de Atribuição de Técnico:
       
       * Para que uma Ordem de Serviço saia de 'PEDING' é obrigatório atribuir um Técnico (Usuário com role TECHNICIAN) Responsável a ela.
       * Uma vez atribuido não será possível trocar Técnico responsável, somente um Usuário com permições do role 'MANAGER'.
      
       ### Regras para atualização de status por Perfil:
        * **Técnico (TECHNICAL)**: Tem autonomia para atulizar todos status exceto ENTREGUE (DELIVERED) e CANCELADO (CANCELED).
        * **Atendente (ATTENDANT):** Tem permissão para atualizar para ENTREGUE ('DELIVERED') se o status já estiver como COMPLETADO (COMPLETED) e também para CANCELADO (CANCELED) caso o status ainda esteja como PEDENTE (PEDING).
        * **GERENTE (MANAGER):** Possui privilégios totais sobre a alterações de status. Mas não pode atualizar se essa Ordem de Serviço j constar comó ENTREGUE (DELIVERED) e com pagamento finalizado.

        * **Completado (COMPLETED):** Ao completar uma Ordem de Serviço automaticamente é gerado um Pagamento já somando todos os gastos e mão de obra realizada.

        ### Validação antes de alterações:
        * **Tarefas e Pagamentos:** O sistema impede a conclusão da OS se houver tarefas pendentes e bloqueia a entrega ao cliente caso as validações de pagamento não sejam atendidas.
        * **Histórico Automático:** Qualquer alteração de status gera um registro de auditoria imutável no histórico da Ordem de Serviço.
        * **Rotina Financeira:** O endpoint aciona automaticamente o processamento e atualização de fluxos de pagamento atrelados à OS.
      requestBody: 
        content: 
          application/json: 
            schema:
              $ref: '#/components/schemas/ServiceOrderUpdateDTO'
      responses: 
        '200':
          description: Ordem de Serviço atualizada com sucesso.
          content: 
            application/json: 
              schema: 
                $ref: '#/components/schemas/ServiceOrderFullDTO'
        '404':
            description: Ordem de Serviço não encontrada com ID informado.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetail'
        '400':
            description: Corpo da requisição mal formatado, ausente ou contendo campos com valores inválidos.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetail' 
        '422':
            description: Os dados estão corretos, mas por Regras de Negócios da API a ação não foi permitida.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetail'   
        '403':
            description: Acesso negado. Apenas usuários com permissão.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetail'
...
  /api/v2/service-orders:
     get: 
       parameters: 
         - name: status 
           in: query
           schema:
            type: string 
           description: Status da Ordem Serviço 
         - name: category
           in: query 
           schema:
            type: string 
           description: Categoria de Dispisitivo
         - name: page 
           in: query
           schema:
            type: number 
           description: Número da página (primeira começa em 0)
         - name: size 
           in: query 
           schema:
            type: number 
           description: Número de elementos por página
       summary: Busca todas Ordem de serviços
       description: |
        Endpoint busca todos as Ordem de Serviço retornando em Paginação, sendo possível filtrar por status e categoria. Por padrão se não for passado nenhum parametro na query, ela retorna somentes as OS que ainda estão em andamento pela ordem que foi criada. Caso seja filtrado por algum status que indica que a OS já foi finalizada, será retornado todas começando pelas recém finalizadas. O mesmo vale para Categoria, por padrão retorna todas sem nenhum filtro.
       tags:
        - Service Orders
       responses:
        '200': 
          description: Busca realizada com sucesso.
          content: 
            application/json: 
              schema:
                $ref: '#/components/schemas/PageServiceOrderDetailDTO'
        '400':
            description: Corpo da requisição mal formatado, ausente ou contendo campos com valores inválidos.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetail' 
        '403':
            description: Acesso negado. Apenas usuários com permissão.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetail'
...
  /api/v2/service-order-items/service-order/{id}:
    parameters: 
      - name: id 
        in: path
        schema:
          type: number 
        description: ID da Ordem Serviço 
        required: true
    get:
      summary: Busca lista de Items (Paças) usadas por Ordem Serviço 
      tags:
        - Service Order Items 
      responses:
        '200':
          description: Busca realizada com sucesso.
          content:
            application/json: 
              schema:
                $ref: '#/components/schemas/ListServiceOrderItemFull'
        '404':
          description: Ordem Serviço não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '403':
          description: Usuário sem permissão.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '400':
            description: Corpo da requisição mal formatado, ausente ou contendo campos com valores inválidos.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetail'
...
  /api/v2/service-order/{id}/payment/history:
    parameters: 
      - name: id
        in: path 
        schema:
          type: number 
        required: true 
    get: 
      tags:
        - History Service Order
      summary: Busca histórico de atualizações de Pagamento por ID da Ordem Serviço 
      responses:
        '200': 
          description: Busca realizada com sucesso
          content: 
            application/json: 
              schema:
                $ref: '#/components/schemas/PagePaymentsHistoryFull' 
        '403':
            description: Usuário sem permissão.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Não foi encontrado Ordem de Serviço informado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Token não informado ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v2/service-order/{id}/history/updates:
    parameters: 
      - name: id
        in: path 
        schema:
          type: number 
        required: true 
    get: 
      tags:
        - History Service Order
      summary: Busca histórico de atualizações da Ordem Serviço pelo ID da OS
      responses:
        '200': 
          description: Busca realizada com sucesso
          content: 
            application/json: 
              schema:
                $ref: '#/components/schemas/PageServiceOrderHistoryFull' 
        '403':
            description: Usuário sem permissão.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Não foi encontrado Ordem de Serviço informado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Token não informado ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'     
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Insira o token JWT no formato "Bearer {token}"
  schemas:
    PageMetadata:
      type: object
      description: Componente unificado contendo os metadados da paginação do Spring Boot.
      properties:
        totalPages:
          type: integer
          description: Quantidade total de páginas resultantes.
          example: 3
        number:
          type: integer
          description: Número da página atual(começa por 0)
          example: 0
        size:
          type: integer
          description: Limite máximo de elementos exibidos por página.
          example: 10
        totalElements:
          type: integer
          format: int64
          description: Quantidade total de registros em toda a base de dados.
          example: 25
    ProblemDetail:
      type: object
      description: Padrão oficial RFC 7807 para mapeamento e retorno de erros de API adotado pelo Spring Boot.
      required:
        - title
        - status
        - detail
        - instance
      properties:
        title:
          type: string
          description: Um resumo curto sobre o tipo do problema.
          example: "Bad Request"
        status:
          type: integer
          description: O código de status HTTP retornado na requisição.
          example: 400
        detail:
          type: string
          description: Uma explicação específica sobre a ocorrência do problema.
          example: "O corpo da requisição é inválido ou atributos não atendem aos critérios."
        instance:
          type: string
          description: URI de referência que identifica a chamada e o endpoint que gerou o erro.
          example: "/api/v2/users"
        errors:
          type: array
          description: Um array opcional contendo o detalhamento de múltiplos erros de validação(Erro de @Valid).
          items:
            type: object
            properties:
              field:
                type: string
                description: Nome do atributo enviado no corpo que falhou na validação.
                example: "costPrice"
              message:
                type: string
                description: O motivo exato da falha do atributo validado.
                example: "O valor não pode ser menor que zero"
...
    ServiceOrderDetailDTO:
      type: object
      properties:
        id:
          type: integer
          description: "Identificador único da ordem de serviço"
          example: 1
        nameClient:
          type: string
          description: "Nome completo do cliente"
          example: "Roberto Silva"
        cpfClient:
          type: string
          description: "CPF do cliente"
          example: "12345678900"
        phone:
          type: string
          description: "Telefone de contato principal do cliente"
          example: "61999998888"
        whatsapp:
          type: string
          description: "Número de WhatsApp do cliente para envio de atualizações"
          example: "61999998888"
        nameEmployee:
          type: string
          description: "Nome do funcionário que registrou o atendimento"
          example: "Ednei Gonzaga"
        device:
          type: string
          description: "Modelo do dispositivo registrado na ordem"
          example: "Samsung Galaxy A54"
        category:
          type: string
          description: "Categoria de classificação do dispositivo"
          example: "SMARTPHONE"
        problemDescription:
          type: string
          description: "Descrição detalhada do defeito relatado"
          example: "Display OLED danificado, necessita de troca do painel e verificação da taxa de atualização de 120Hz."
        identificationCode:
          type: string
          description: "Código alfanumérico único para rastreio rápido da OS"
          example: "123456789"
        status:
          type: string
          description: "Situação atual do andamento do serviço"
          example: "EM_ANDAMENTO"
        dateTimeStart:
          type: string
          description: "Data e hora em que o serviço foi iniciado"
          example: "2026-07-17T10:30:00"
        dateTimeCompleted:
          type: string
          description: "Data e hora em que o serviço foi finalizado"
          example: "2026-07-18T16:00:00"
        dateTimeUpdateStatus:
          type: string
          description: "Data e hora da última alteração de status da OS"
          example: "2026-07-17T14:15:00"
        userTechnical:
          type: integer
          description: "ID interno do técnico responsável pela execução do serviço"
          example: 3
        serviceRequest:
          type: integer
          description: "ID da requisição de orçamento/serviço que originou esta OS"
          example: 42
    ServiceOrderUpdateDTO:
      type: object
      properties:
        userTechnical:
          type: integer
          description: "ID interno do técnico que será atribuído ou atualizado na ordem de serviço"
          example: 3
        status:
          type: string
          description: "Novo status a ser definido para a ordem de serviço"
          example: "EM_ANDAMENTO"
    ServiceOrderFullDTO:
      type: object
      properties:
        id:
          type: integer
          description: "Identificador único da ordem de serviço"
          example: 1
        identificationCode:
          type: string
          description: "Código  único para rastreio rápido da OS"
          example: "123456789"
        status:
          type: string
          description: "Situação atual do andamento do serviço"
          example: "EM_ANDAMENTO"
        dateTimeStart:
          type: string
          description: "Data e hora em que o serviço foi iniciado"
          example: "2026-07-17T10:30:00"
        dateTimeCompleted:
          type: string
          description: "Data e hora em que o serviço foi finalizado"
          example: "2026-07-18T16:00:00"
        dateTimeUpdateStatus:
          type: string
          description: "Data e hora da última alteração de status da OS"
          example: "2026-07-17T14:15:00"
        serviceRequest:
          type: integer
          description: "ID da requisição serviço que originou esta OS"
          example: 42
        userTechnical:
          type: integer
          description: "ID interno do técnico responsável pela execução do serviço"
          example: 3
    PageServiceOrderDetailDTO:
      type: object 
      allOf:
        - type: object
          properties:
            content:
              type: array
              items:
                  $ref: '#/components/schemas/ServiceOrderDetailDTO'
        - $ref: '#/components/schemas/PageMetadata'
    ServiceOrderItemFullDTO:
      type: object
      properties: 
        id:
          type: number 
          description: ID do Item de Ordem Serviço
          example: 3
        serviceOrder:
          type: number 
          description: ID da Orderm de Serviço 
          example: 5
        part:
          type: number 
          description: ID da peça ultilizada 
          example: 6
        namePart: 
          type: string 
          description: Nome da peça 
          example: SSD NVME 500GB
        quantity:
          type: string 
          description: Quantidade usada da peça 
          example: 2
        unitPrice:
          type: number 
          description: Valor pago na peça 
          example: 120.0
        subTotal:
          type: number 
          description: Valor Total gasto com as peças
          example: 240.0
...
    ListServiceOrderItemFull:
       type: array 
       items:
        $ref: '#/components/schemas/ServiceOrderItemFullDTO'
...
    PaymentsHistoryFullDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único do histórico
        paymentId:
          type: integer
          format: int64
          description: Identificador do pagamento associado
        userId:
          type: integer
          format: int64
          description: Identificador do usuário que realizou a ação
        oldStatus:
          type: string
          description: Status anterior do pagamento
          example: "PENDING"
        newStatus:
          type: string
          description: Status novo do pagamento
          example: "CANCELED"
        transactionAmount:
          type: number
          format: double
          description: Valor da transação registrada
          example: 150.75
        notes:
          type: string
          description: Observações adicionais sobre a mudança de status
          example: "Pagamento aprovado manualmente pelo administrador"
        createdAt:
          type: string
          format: date-time
          description: Data e hora em que o histórico foi criado
    PagePaymentsHistoryFull:
      allOf: 
        - type: object
          properties: 
            content:
              type: array 
              items:
                $ref: '#/components/schemas/PaymentsHistoryFullDTO'
        - $ref: '#/components/schemas/PageMetadata'
    ServiceOrderHistoryFullDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único do histórico
        serviceOrderId:
          type: integer
          format: int64
          description: Identificador da ordem de serviço
        userId:
          type: integer
          format: int64
          description: Identificador do usuário que realizou a ação
        oldStatus:
          type: string
          description: Status anterior da ordem de serviço
          example: "PEDING"
        newStatus:
          type: string
          description: Status novo da ordem de serviço
          example: "COMPLETED"
        notes:
          type: string
          description: Observações adicionais sobre a mudança de status
          example: "Finalizado após revisão técnica"
        createdAt:
          type: string
          format: date-time
          description: Data e hora em que o histórico foi criado
    PageServiceOrderHistoryFull:
      allOf:
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ServiceOrderHistoryFullDTO'
        - $ref: '#/components/schemas/PageMetadata'